Like-icons
Updated the icons to follow our latest SVF-code conventions. 2 new
icons with the colours needed for the liking action.

OK, here come's the next problem with the icon naming-convention (i was
waiting for that)! These coloured icons get the same name as usually
the standard blue icons would get. So name-conflicts are just a
question of time … :(